refactor(react-static): extract emitFileWriteMetrics shared by both renderers
renderPages.ts and renderPagesBatched.ts inlined the same ~90-line
file.write.done metrics block (rebuild html / rsc-full / rsc-headless render
metrics from the on-disk file data), differing only in the name of the results
Map. Extract it to emitFileWriteMetrics(event, route, routeResults, options);
both full paths now call it.

The html-only skip-branch in renderPages is left inline on purpose — it emits a
subset by design, so reusing the full helper there would add rsc/rsc-full
metrics (a behavior change).

Behavior preserved: test:metrics asserts the emitted metric values (type,
fileSize, chunks, processingTime, streamMetrics.duration) and passes; server +
client gates green (622 / 180). Net -220 lines.
